+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Changed
+
+- [store] remove dep on `@cross/env`
+
+### Added
+
+- [store-node-fs] Node.js compatible file storage
+
 ## [0.3.0]
 
 ### Changed